refactor: _libをquery/とmutation/に分離
- _lib/query/配下にfetch関数を集約
- _lib/mutation/配下にmutation関数を集約
- useFollowMutationsからfollowUser/unfollowUserを分離
- ディレクトリ構造で責務を明確化（query=読み取り, mutation=書き込み）
- すべてのインポートパスを更新
