+++ frontend/src/app/(protected)/users/_hooks/useFollowStatus.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { queryKeys } from '@/constants/queryKeys';
-import { fetchFollowStatus } from '@/app/(protected)/users/_lib/fetchFollowStatus';
+import { fetchFollowStatus } from '@/app/(protected)/users/_lib/query/fetchFollowStatus';
 
 export function useFollowStatus(userId: string) {
   return useQuery({
