+++ frontend/src/app/(protected)/users/_hooks/useFollowing.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { queryKeys } from '@/constants/queryKeys';
-import { fetchFollowing } from '@/app/(protected)/users/_lib/fetchFollowing';
+import { fetchFollowing } from '@/app/(protected)/users/_lib/query/fetchFollowing';
 
 export function useFollowing(userId: string) {
   return useQuery({
