+++ frontend/src/app/(protected)/users/_hooks/useUser.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { queryKeys } from '@/constants/queryKeys';
-import { fetchUser } from '@/app/(protected)/users/_lib/fetchUser';
+import { fetchUser } from '@/app/(protected)/users/_lib/query/fetchUser';
 
 export function useUser(id: string) {
   return useQuery({
