+++ frontend/src/app/(protected)/users/_hooks/useUserBooks.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { queryKeys } from '@/constants/queryKeys';
-import { fetchUserBooks } from '@/app/(protected)/users/_lib/fetchUserBooks';
+import { fetchUserBooks } from '@/app/(protected)/users/_lib/query/fetchUserBooks';
 
 export function useUserBooks(id: string) {
   return useQuery({
