+++ frontend/src/app/(protected)/users/_hooks/useUserLists.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { queryKeys } from '@/constants/queryKeys';
-import { fetchUserLists } from '@/app/(protected)/users/_lib/fetchUserLists';
+import { fetchUserLists } from '@/app/(protected)/users/_lib/query/fetchUserLists';
 
 export function useUserLists(id: string) {
   return useQuery({
