+++ frontend/src/app/(protected)/users/_lib/mutation/followUser.ts
@@ -1,0 +1,20 @@
+interface FollowResponse {
+  message: string;
+}
+
+interface FollowError {
+  error: string | string[];
+}
+
+export async function followUser(userId: string): Promise<FollowResponse> {
+  const response = await fetch(`/api/users/${userId}/follow`, {
+    method: 'POST',
+  });
+
+  if (!response.ok) {
+    const error: FollowError = await response.json();
+    throw new Error(Array.isArray(error.error) ? error.error.join(', ') : error.error);
+  }
+
+  return response.json();
+}
