+++ frontend/src/app/(protected)/users/_lib/query/fetchFollowing.ts
